+++ src/components/ProjectItem.vue
@@ -1,5 +1,3 @@
-<!-- ProjactItem.vue -->
-
 <template>
   <div
     class="project-item"
@@ -15,55 +13,8 @@
 </template>
 
 <script>
-export default {
-  props: {
-    project: {
-      type: Object,
-      required: true,
-    },
-  },
-  data() {
-    return {
-      showDelete: false,
-    };
-  },
-  methods: {
-    showDeleteButton() {
-      this.showDelete = true;
-    },
-    hideDeleteButton() {
-      this.showDelete = false;
-    },
-    deleteProject() {
-      this.$emit("delete-project", this.project.name);
-    },
-  },
-};
+import ProjectItem from "../scripts/ProjectItem";
+export default ProjectItem;
 </script>
 
-<style scoped>
-.project-item {
-  position: relative;
-  padding: 10px;
-  background-color: #f9f9f9;
-  border-radius: 5px;
-  box-shadow: 0 0 5px rgba(0, 0, 0, 0.1);
-  margin-bottom: 10px;
-}
-
-.delete-button {
-  position: absolute;
-  top: 5px;
-  right: 5px;
-  background-color: red;
-  color: white;
-  border: none;
-  border-radius: 50%;
-  width: 20px;
-  height: 20px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-}
-</style>
+<style scoped src="../styles/ProjectItem.css"></style>
