+++ src/scripts/AdvancedSearch.js
@@ -1,197 +1,3 @@
-<template>
-  <div class="advanced-search">
-    <div class="ads-top">
-      <div class="dropdown">
-        <button class="dropbtn" @click="toggleDropdown">選擇分析功能</button>
-        <div class="dropdown-content" v-show="dropdownOpen">
-          <a
-            @click="
-              hidden_btn();
-              show_search(1);
-            "
-            >關鍵字</a
-          >
-          <a
-            @click="
-              hidden_btn();
-              show_search(3);
-            "
-            >關鍵字成長</a
-          >
-          <a
-            @click="
-              hidden_btn();
-              show_search(4);
-            "
-            >作者</a
-          >
-          <a
-            @click="
-              hidden_btn();
-              show_search(5);
-            "
-            >引用次數</a
-          >
-          <a
-            @click="
-              hidden_btn();
-              show_search(6);
-            "
-            >研究領域</a
-          >
-          <a
-            @click="
-              hidden_btn();
-              show_search(8);
-            "
-            >領域成長</a
-          >
-        </div>
-      </div>
-      <!-- <button class="btn-option" @click="hidden_btn()" v-if="!btn_show">
-        返回
-      </button> -->
-
-      <div class="search-group">
-        <!-- 關鍵字，年份區間 -->
-        <div class="year-keyword search-item" v-if="search_block === 1">
-          <p>請輸入年份區間，系統會顯示該區間之關鍵字及出現次數</p>
-          <label>
-            開始年:
-            <input type="number" v-model="startYear" class="small-input" />
-          </label>
-          <label>
-            結束年:
-            <input type="number" v-model="endYear" class="small-input" />
-          </label>
-          <label>
-            最少出現次數(下限):
-            <input type="number" v-model="lower_limit" class="small-input" />
-          </label>
-          <div>
-            <input
-              type="button"
-              value="開始分析"
-              @click="startAnalysis_keywordYear()"
-            />
-          </div>
-        </div>
-
-        <!-- 單一關鍵字 -->
-        <div class="single-keyword search-item" v-if="search_block === 3">
-          <p>根據關鍵字做查詢，可觀察該關鍵字每年的成長趨勢</p>
-          <label>
-            輸入關鍵字
-            <input
-              type="text"
-              v-model="single_key"
-              @keyup.enter="startAnalysis_singleKeyword()"
-              class="small-input"
-            />
-          </label>
-          <div>
-            <input
-              type="button"
-              value="開始分析"
-              @click="startAnalysis_singleKeyword()"
-            />
-          </div>
-        </div>
-
-        <!-- 作者 年份區間 -->
-        <div class="year-author search-item" v-if="search_block === 4">
-          <p>根據年份區間對作者做分析（看年份區間內作者發表了幾篇）</p>
-          <label>
-            開始年:
-            <input type="number" v-model="startYear" class="small-input" />
-          </label>
-          <label>
-            結束年:
-            <input type="number" v-model="endYear" class="small-input" />
-          </label>
-          <label>
-            最少出現次數(下限):
-            <input type="number" v-model="lower_limit" class="small-input" />
-          </label>
-          <div>
-            <input
-              type="button"
-              value="開始分析"
-              @click="startAnalysis_yearAuthor()"
-            />
-          </div>
-        </div>
-        <!-- 根據引用次數做分析（一併提供標題和作者資訊） -->
-        <div class="author-cite search-item" v-if="search_block === 5">
-          <p>根據引用次數做分析（一併提供標題和作者資訊）</p>
-          <label>
-            被引用最少次數(下限):
-            <input type="number" v-model="lower_limit" class="small-input" />
-          </label>
-          <div>
-            <input
-              type="button"
-              value="開始分析"
-              @click="startAnalysis_author_cite()"
-            />
-          </div>
-        </div>
-
-        <!-- 領域 年份區間 -->
-        <div class="field-year search-item" v-if="search_block === 6">
-          <p>根據年份區間做研究領域分析</p>
-          <label>
-            開始年:
-            <input type="number" v-model="startYear" class="small-input" />
-          </label>
-          <label>
-            結束年:
-            <input type="number" v-model="endYear" class="small-input" />
-          </label>
-          <label>
-            最少出現次數(下限):
-            <input type="number" v-model="lower_limit" class="small-input" />
-          </label>
-          <div>
-            <input
-              type="button"
-              value="開始分析"
-              @click="startAnalysis_fieldYear()"
-            />
-          </div>
-        </div>
-        <!-- 單一領域成長趨勢 -->
-        <div class="single-field search-item" v-if="search_block === 8">
-          <p>根據研究領域做查詢，可觀察該研究領域每年的成長趨勢</p>
-          <label>
-            輸入關鍵字
-            <input
-              type="text"
-              v-model="single_field"
-              @keyup.enter="startAnalysis_singleField()"
-              class="small-input"
-            />
-          </label>
-          <div>
-            <input
-              type="button"
-              value="開始分析"
-              @click="startAnalysis_singleField()"
-            />
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <div class="chart-show">
-      <div id="chart" v-if="showChart"></div>
-      <VueSpinnerHourglass size="40px" color="blue" v-if="waitComp" />
-      <h3 v-if="waitComp">資料分析需要時間，請勿重複點擊開始分析</h3>
-    </div>
-  </div>
-</template>
-
-<script setup>
 import { ref, watch } from "vue";
 import { VueSpinnerHourglass } from "vue3-spinners";
 
@@ -752,12 +558,3 @@ async function drawChart_singleField() {
     chart.draw(data, options);
   });
 }
-</script>
-
-<style scoped src="../styles/AdvancedSearch.css"></style>
-<!-- . -->
-
-<!--  
-import AdvancedSearch from "../scripts/AdvancedSearch.js";
-export default AdvancedSearch;
--->
