+++ src/styles/AdvancedSearch.css
@@ -1,0 +1,178 @@
+.advanced-search {
+  width: 70%;
+  margin: 20px auto;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: center;
+  align-content: flex-start;
+}
+.ads-top {
+  width: 100%;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+}
+.dropdown {
+  position: relative;
+  display: inline-block;
+  margin-bottom: 195px;
+}
+
+.dropbtn {
+  width: 170%;
+  background-color: #333;
+  color: white;
+  padding: 10px 20px;
+  border: none;
+  border-radius: 10px;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+}
+
+.dropbtn:hover {
+  background-color: #555;
+}
+
+.dropdown-content {
+  width: 150%;
+  display: none;
+  position: absolute;
+  background-color: #f9f9f9;
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.2);
+  border-radius: 10px;
+  z-index: 1;
+}
+
+.dropdown-content a {
+  color: black;
+  padding: 12px 16px;
+  text-decoration: none;
+  display: block;
+  border-bottom: 1px solid #ddd;
+}
+
+.dropdown-content a:hover {
+  background-color: #f1f1f1;
+}
+
+.dropdown:hover .dropdown-content {
+  display: block;
+}
+
+.btn-option {
+  background-color: #333;
+  color: white;
+  padding: 10px 20px;
+  margin: 10px;
+  border: none;
+  border-radius: 10px;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+}
+.btn-group {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: space-around;
+  margin-bottom: 20px;
+  width: 100%;
+}
+
+.btn-option:hover {
+  background-color: #555;
+}
+
+.search-group {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.search-item {
+  margin-bottom: 20px;
+  padding: 20px;
+  background-color: #f9f9f9;
+  border-radius: 10px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  width: 80%;
+  text-align: center;
+}
+
+.search-item p {
+  font-size: 24px;
+  font-weight: bold;
+}
+
+.search-item label {
+  display: block;
+  margin: 10px 0;
+}
+
+.search-item input[type="number"],
+.search-item input[type="text"] {
+  padding: 10px;
+  margin-top: 10px;
+  width: calc(100% - 22px);
+  border: 1px solid #ddd;
+  border-radius: 10px;
+}
+
+.search-item button {
+  background-color: #333;
+  color: white;
+  padding: 10px 20px;
+  margin-top: 20px;
+  border: none;
+  border-radius: 10px;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+}
+
+.search-item button:hover {
+  background-color: #555;
+}
+.checkbox-row {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+.chart-show {
+  width: 100%;
+  height: 400px;
+  max-width: 1200px;
+  margin-top: 20px;
+  display: flex;
+  justify-content: center;
+}
+#chart {
+  width: 100%;
+  height: 100%;
+  margin: 0 auto;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.small-input {
+  width: 60px;
+  padding: 5px;
+  border: 1px solid #d1d5db;
+  border-radius: 5px;
+  margin-right: 10px;
+}
+
+.search-item input[type="button"] {
+  background-color: #3b82f6;
+  color: white;
+  padding: 5px 10px;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+}
+
+.search-item input[type="button"]:hover {
+  background-color: #2563eb;
+}
