+++ src/styles/Projectlist.css
@@ -1,0 +1,151 @@
+.project-list {
+  width: 250px;
+  height: 100%;
+  border-right: 1px solid #ddd;
+  padding: 10px;
+  box-sizing: border-box;
+  transition: width 0.3s;
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  background-color: #f7f7f7;
+  border-radius: 10px;
+}
+
+ul {
+  list-style-type: none;
+  padding: 0;
+  margin: 0;
+}
+
+li {
+  padding: 10px;
+  border-bottom: 1px solid #ddd;
+  border-radius: 10px;
+  position: relative;
+}
+
+li:hover {
+  background-color: #e0e0e0;
+  cursor: pointer;
+}
+
+button {
+  margin-top: 20px;
+  padding: 10px;
+  width: 100%;
+  background-color: #333;
+  color: white;
+  border: none;
+  border-radius: 10px;
+  cursor: pointer;
+}
+
+button:hover {
+  background-color: #555;
+}
+
+.delete-button {
+  background-color: red;
+  color: white;
+  border: none;
+  border-radius: 50%;
+  width: 20px;
+  height: 20px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+  position: absolute;
+  top: 5px;
+  right: 5px;
+}
+
+.toggle-button {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 20px;
+  height: 40px;
+  background-color: #333;
+  color: white;
+  cursor: pointer;
+  position: absolute;
+  right: -10px;
+  top: 50%;
+  transform: translateY(-50%);
+  transition: right 0.3s;
+  border: none;
+  border-radius: 5px;
+}
+
+.toggle-button span {
+  font-size: 14px;
+}
+
+.collapsed .project-list {
+  width: 20px;
+  overflow: hidden;
+}
+
+.collapsed ul,
+.collapsed button {
+  display: none;
+}
+
+.collapsed .toggle-button {
+  right: 0;
+  background-color: #555;
+}
+
+.active {
+  background-color: #ccc;
+  border-radius: 10px;
+}
+
+.modal {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: fixed;
+  z-index: 1;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  overflow: auto;
+  background-color: rgba(0, 0, 0, 0.5);
+}
+
+.modal-content {
+  background-color: #fefefe;
+  padding: 20px;
+  border: 1px solid #888;
+  border-radius: 10px;
+  width: 300px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+}
+
+.close {
+  color: #aaa;
+  float: right;
+  font-size: 28px;
+  font-weight: bold;
+  cursor: pointer;
+}
+
+.close:hover,
+.close:focus {
+  color: black;
+  text-decoration: none;
+  cursor: pointer;
+}
+
+.large-input {
+  width: 80%;
+  padding: 5px;
+  font-size: 16px;
+  border: 1px solid #ddd;
+  border-radius: 10px;
+  margin-bottom: 20px;
+}
